requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-substrate-interface==1.1.0
async-timeout==5.0.1
asyncstdlib==3.13.1
attrs==25.3.0
base58==2.1.1
billiard==4.2.1
bittensor==9.3.0
bittensor-commit-reveal==0.3.1
bittensor-wallet==3.0.8
bt-decode==0.6.0
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==43.0.3
cytoolz==1.0.1
decorator==5.2.1
dnspython==2.7.0
email_validator==2.2.0
eth-hash==0.7.1
eth-typing==5.2.0
eth-utils==2.2.2
exceptiongroup==1.2.2
fastapi==0.110.3
fastapi-cli==0.0.7
flower==2.0.1
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanize==4.12.2
idna==3.10
Jinja2==3.1.6
kombu==5.5.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
msgpack==1.1.0
msgpack-numpy-opentensor==0.5.0
multidict==6.4.2
munch==2.5.0
nest-asyncio==1.6.0
netaddr==1.3.0
numpy==2.0.2
packaging==24.2
password-strength==0.0.3.post2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.1
py==1.11.0
py-bip39-bindings==0.1.11
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-statemachine==2.5.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
retry==0.9.2
rich==14.0.0
rich-toolkit==0.14.1
scalecodec==1.2.11
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
toml==0.10.0
toolz==1.0.0
tornado==6.4.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
xxhash==3.5.0
yarl==1.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
